Implement Trace Generation Framework
This implements a relatively straightforward trace generation framework.
Specifically, it generates all possible trace data upto a given length
for a given set of constraints.  Then, using a model written in Rust, it
divides that trace data into two piles: the `accepts` and the `rejects`.
The former are those which should be accepted by the constraints, whilst
the latter are those which should be rejected.