Fix #158
This puts through a fix for validation of imported values when `native`
fields are being used.  The issue is that validation was being
performed _after_ conversion into the native field (when this was
selected via the `-N` cli switch).  Unfortunately, validation at that
point cannot work because every field value is assumed to require `254`
bits.