Second Fix for #158
This puts through alternative fix for #158 which works for other magma
types as well.  This does not attempt to bypass the magma validation.
Instead, it changes Magma validation so that it no longer uses the
`bit_size` of the target value and, instead, checks the target value
against the maximum possible value.  Specifically, for a value `x` of type
`iN` it checks that `x < 2^N`.