Fix Ambiguity in Target/Source Module for lookups
This fixes potential ambiguities in the source or target module for a
lookup constraint.  The essentially is that a single module must be
identified for the source (resp. target) columns of a lookup.  This was
not previously checked.  Furthermore, there was a problem when no such
module was evident from a given column (e.g. if it was a constant
expression).