Partial Fix for #201
This puts in place an almost complete fix for #201.  Specifically,
instead of constructing `ComputedColumn`s for columns, it now does a
very similar process for registers.  This is not the most efficient
approach, as it requires searching for the columns allocated to a given
register.  Then, it determines the padding value for all columns and
checks they match.  However, improving this would require more invasive
changes which are out-of-scope at this time.

One issue remains however.  Specicifically, there appear to be registers
which are not associated with any column.  I'm not sure how this could
arise, or what's going on here.