feat: Support `RON` and `JSON` bin formats (#223)
Previously, the ability to generate `JSON` binfiles was hidden behind a
build feature. Hence, you needed to build the binary specifically for
generating `JSON` binfiles. The primary reason for this is that
supporting both `RON` and `JSON` formats in the same binary required
breaking changes to the format. However, the time has come to address
this and support both formats. At this time, `RON` remains the default
choice --- but this will be deprecated in the very near future.