Replace `raise_ifs` and `do_expand_ifs`
This replaces these functions with `lower_expr()` which more closely
follows the Go implementation.  Currently, this is passing all tests but
needs more work to be sure.  A few points:

* Need to remove the now redundant code.
* Need to better manage optimisation of constant expressions.

The problem with the latter is that we probably need to do the constant
evaluation at the point of an `if` condition.